simplified ConfigurableCountryChoiceType configureOptions not to use setNormalizer
